% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ch3.R
\docType{data}
\name{ch3}
\alias{ch3}
\title{ch3 dataset #'
#' Dataset imported from test_data.csv.
#' #' @format A data frame.
#' @source Original file: test_data.csv}
\format{
An object of class \code{data.frame} with 347 rows and 21 columns.
}
\usage{
ch3
}
\description{
ch3 dataset #'
#' Dataset imported from test_data.csv.
#' #' @format A data frame.
#' @source Original file: test_data.csv
}
\keyword{datasets}
